all actions with grid buttons

+++ src/app/modules/requests/components/request/request.component.ts
@@ -1,7 +1,6 @@
-import { Component, OnInit, Input } from '@angular/core';
+import { Component, OnInit, Inject, ViewChild, ElementRef } from '@angular/core';
 import { FormBuilder, FormGroup, Validators, FormControl, AbstractControl } from '@angular/forms';
 import { throwError } from 'rxjs';
-// import { DatePipe } from '@angular/common';
 
 import { Action } from '../../../../shared/model/action';
 import { Request } from '../../model/request';
@@ -12,6 +11,10 @@ import { RequestStatus } from '../../model/request-status';
 import { DistributionType } from '../../model/distribution-type';
 import { Guid } from '../../../../shared/model/guid';
 import { Validator } from '../../../../shared/core/validator';
+import { ActivatedRoute } from '@angular/router';
+import { MatDialog, MAT_DIALOG_DATA } from '@angular/material/dialog';
+import { ConfirmationComponent } from '../../../dialogs/confirmation.component';
+
 
 @Component({
   selector: 'ms-request',
@@ -19,17 +22,23 @@ import { Validator } from '../../../../shared/core/validator';
   styleUrls: ['./request.component.css']
 })
 export class RequestComponent implements OnInit {
+  constructor(private formBuilder: FormBuilder, private requestService: RequestService, private route: ActivatedRoute,
+    private dialog: MatDialog, @Inject(MAT_DIALOG_DATA) public data: { request: Request, action: Action }) {
+    this.action = data.action;
+    this.request = data.request ? data.request : this.getDefaultRequest();
+  }
 
-  constructor(private formBuilder: FormBuilder, private requestService: RequestService) { }
-
-  @Input() request?: Request;
-  @Input() action: Action = Action.Add;
+  readonly request?: Request;
+  readonly action: Action = Action.Add;
   post: any;
   requestForm: FormGroup;
   validator: Validator;
 
+  AActions = Action;
   PacksgeTypes = PackageType;
-  packsgeTypesKeys: string[] = Object.keys(PackageType).filter(Number);
+  PacksgeTypesValues: string[] = Object.values(PackageType).filter(Number);
+  @ViewChild(ElementRef, { static: false }) selectPackageType: ElementRef;
+
 
   get user() { return this.requestForm.get('user') as FormControl; }
   get email() { return this.requestForm.get('email') as FormControl; }
@@ -43,7 +52,7 @@ export class RequestComponent implements OnInit {
 
   initializeForm() {
     this.requestForm = this.formBuilder.group({
-      user: [null, [Validators.required, Validators.minLength(5), Validators.maxLength(20), this.validateLowerUsername]],
+      user: [null, [Validators.required, Validators.minLength(5), Validators.maxLength(50), this.validateUsername]],
       email: [null, [Validators.required, Validators.email]],
       package: this.formBuilder.group({
         name: [null, [Validators.required, Validators.minLength(5), Validators.maxLength(128)]],
@@ -51,8 +60,7 @@ export class RequestComponent implements OnInit {
       }),
       distribution: [null]
     });
-    if (!this.request)
-      this.request = this.getDefaultRequest();
+
     this.requestForm.patchValue(this.request);
     this.validator = new Validator(this.requestForm);
   }
@@ -62,27 +70,45 @@ export class RequestComponent implements OnInit {
       return;
 
     Object.keys(this.requestForm.value).forEach(key => this.request[key] = this.requestForm.value[key]); // bind form data
-    // this.request.submittedOn =  this.datePipe.transform(this.request.submittedOn, 'dd-MM-yyyy');
-
     switch (this.action) {
       case Action.Add:
-        this.requestService.add(this.request);
+        this.requestService.add(this.request).subscribe(req => {
+          this.requestService.getAll();
+        });
         break;
       case Action.Modify:
         this.requestService.update(this.request);
         break;
       case Action.Delete:
-        this.requestService.delete(this.request.id);
+        const dialogRef = this.dialog.open(ConfirmationComponent, {
+          data: {
+            message: `Are you sure to delete package ${this.request.package.name}?`,
+            buttonText: {
+              ok: 'Yes',
+              cancel: 'No'
+            }
+          }
+        });
+
+        dialogRef.afterClosed().subscribe((confirmed: boolean) => {
+          if (confirmed)
+            this.requestService.delete(this.request.id);
+        });
+
+
+        // if (confirm('Are you sure to delete package ' + this.request.package.name))
+        //   this.requestService.delete(this.request.id);
         break;
       default:
         throwError('Action required');
         break;
     }
+    this.requestService.dialogRef.close(this.request.id);
   }
 
-  private validateLowerUsername(control: AbstractControl) {
+  private validateUsername(control: AbstractControl) {
     const username: string = control.value;
-    return (username && username.toLowerCase() !== username) ? { 'invalidValue': 'must be all lowercase' } : null;
+    return (!username /*&& username.toLowerCase() !== username*/) ? { 'invalidValue': 'must be all lowercase' } : null;
   }
 
   private getDefaultRequest(): Request {
